update library and function calls
In middleman.sh, the user is instructed to fork the upstream project, rename it  and then the script will clone it from their repository. In this way the user gets their own copy to modify. Upstream changes can still be merged in the future if the user chooses to do so. No need to create a separate branch for the user's middleman site since they will be customizing their project. It will diverge from the upstream one.

Changes to be committed:

	modified:   includes/git.lib
	modified:   scripts/middleman.sh